+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * `colsconfig-dialog-content` component:
     * Fix column names saving in export excel mode.
     * Fix usersettings apply, when query params doesn't change.
+* `olv-toolbar` component:
+    * Fix usersettings copy dialog selectors.
 * `flexberry-file` component:
     * Fix mobile file component menu.
 * `flexberry-menu` component: